fix: unify code into single page template utils file
