+++ src/actions/sponsor-pages-actions.js
@@ -22,20 +22,14 @@ import {
   escapeFilterValue
 } from "openstack-uicore-foundation/lib/utils/actions";
 import T from "i18n-react/dist/i18n-react";
-import moment from "moment-timezone";
-import {
-  getAccessTokenSafely,
-  normalizeSelectAllField
-} from "../utils/methods";
+import { getAccessTokenSafely } from "../utils/methods";
 import { snackbarErrorHandler, snackbarSuccessHandler } from "./base-actions";
 import {
   DEFAULT_CURRENT_PAGE,
   DEFAULT_ORDER_DIR,
-  DEFAULT_PER_PAGE,
-  PAGE_MODULES_DOWNLOAD,
-  PAGE_MODULES_MEDIA_TYPES,
-  PAGES_MODULE_KINDS
+  DEFAULT_PER_PAGE
 } from "../utils/constants";
+import { normalizePageTemplateModules } from "../utils/page-template";
 
 export const GLOBAL_PAGE_CLONED = "GLOBAL_PAGE_CLONED";
 export const RESET_EDIT_PAGE = "RESET_EDIT_PAGE";
@@ -568,41 +562,10 @@ const normalizeSponsorCustomPage = (entity, summitTZ) => {
     normalizedEntity.allowed_add_ons = entity.allowed_add_ons.map((e) => e.id);
   }
 
-  normalizedEntity.modules = entity.modules.map((module) => {
-    const normalizedModule = { ...module };
-
-    if (module.kind === PAGES_MODULE_KINDS.MEDIA) {
-      if (module.upload_deadline) {
-        normalizedModule.upload_deadline = moment
-          .tz(module.upload_deadline, summitTZ)
-          .unix();
-      }
-
-      if (module.file_type_id) {
-        normalizedModule.file_type_id =
-          module.file_type_id?.value || module.file_type_id;
-      }
-
-      if (module.type === PAGE_MODULES_MEDIA_TYPES.INPUT) {
-        delete normalizedModule.file_type_id;
-        delete normalizedModule.max_file_size;
-      }
-    }
-
-    if (module.kind === PAGES_MODULE_KINDS.DOCUMENT) {
-      if (module.type === PAGE_MODULES_DOWNLOAD.FILE) {
-        normalizedModule.file = module.file?.[0] || null;
-        delete normalizedModule.external_url;
-      } else {
-        delete normalizedModule.file;
-        delete normalizedModule.file_id;
-      }
-    }
-
-    delete normalizedModule._tempId;
-
-    return normalizedModule;
-  });
+  normalizedEntity.modules = normalizePageTemplateModules(
+    entity.modules,
+    summitTZ
+  );
 
   return normalizedEntity;
 };
